pyproject.toml
[tool.poetry]
name = "image-to-ascii"
version = "0.1.0"
description = ""
authors = ["Ajay Ratnam <ajratnam.dev@gmail.com>"]
readme = "README.md"
packages = [{include = "image_to_ascii"}]

[tool.poetry.dependencies]
python = "^3.10"
pillow = "^10.0.0"
numpy = "^1.25.0"
requests = "^2.31.0"
sty = "^1.0.4"


[tool.poetry.group.dev.dependencies]
mkdocs-material = "^9.1.21"
mkdocstrings-python = "^1.2.1"
pytest = "^7.4.0"
coverage = "^7.2.7"
pytesseract = "^0.3.10"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
